+++ modules/gsuite_group/main.tf
@@ -15,13 +15,15 @@
  */
 
 locals {
-  domain = "${var.domain != "" ? var.domain : data.google_organization.org.domain}"
-  email  = "${format("%s@%s", var.name, local.domain)}"
+  domain_list = "${concat(data.google_organization.org.*.domain, list("dummy"))}"
+  domain      = "${var.domain == "" ? element(local.domain_list, 0) : var.domain}"
+  email       = "${format("%s@%s", var.name, local.domain)}"
 }
 
 /*****************************************
   Organization info retrieval
  *****************************************/
 data "google_organization" "org" {
+  count        = "${var.domain == "" ? 1 : 0}"
   organization = "${var.org_id}"
 }